pyproject.toml
[project]
name = "youtube-to-mp3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pytube>=15.0.0",
    "yt-dlp>=2025.12.8",
]
